+++ test/common.dart
@@ -1,5 +1,5 @@
 const PseudoApplication pseudoApplication =
-    const PseudoApplication('foo', 'bar', 'http://foo.bar/baz');
+    PseudoApplication('foo', 'bar', 'http://foo.bar/baz');
 
 class PseudoApplication {
   final String id, secret, redirectUri;
@@ -7,10 +7,10 @@ class PseudoApplication {
   const PseudoApplication(this.id, this.secret, this.redirectUri);
 }
 
-const List<PseudoUser> pseudoUsers = const [
-  const PseudoUser(username: 'foo', password: 'bar'),
-  const PseudoUser(username: 'michael', password: 'jackson'),
-  const PseudoUser(username: 'jon', password: 'skeet'),
+const List<PseudoUser> pseudoUsers = [
+  PseudoUser(username: 'foo', password: 'bar'),
+  PseudoUser(username: 'michael', password: 'jackson'),
+  PseudoUser(username: 'jon', password: 'skeet'),
 ];
 
 class PseudoUser {